Refactor source into separate modules
