+++ addon/utils/validators/index.js
@@ -1,0 +1,35 @@
+import Ember from 'ember'
+const {assign} = Ember
+
+import array from './array'
+import bool from './bool'
+import emberObject from './ember-object'
+import func from './func'
+import instanceOf from './instance-of'
+import nullFn from './null'
+import number from './number'
+import object from './object'
+import oneOf from './one-of'
+import oneOfType from './one-of-type'
+import shape from './shape'
+import string from './string'
+
+const validators = {
+  array,
+  bool,
+  EmberObject: emberObject,
+  func,
+  instanceOf,
+  null: nullFn,
+  number,
+  object,
+  oneOf,
+  string
+}
+
+assign(validators, {
+  oneOfType: oneOfType.bind(this, validators),
+  shape: shape.bind(this, validators)
+})
+
+export default validators
